+++ lib/flutter_share_me.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/services.dart';
+
 import 'file_type.dart';
 
 //export file type enum
@@ -14,6 +15,7 @@ class FlutterShareMe {
   static const String _methodTwitter = 'twitter_share';
   static const String _methodInstagramShare = 'instagram_share';
   static const String _methodSystemShare = 'system_share';
+  static const String _methodTelegramShare = 'telegram_share';
 
   ///share to WhatsApp
   /// [imagePath] is local image
@@ -21,9 +23,7 @@ class FlutterShareMe {
   /// For ios
   /// If include image then text params will be ingored as there is no current way in IOS share both at the same.
   Future<String?> shareToWhatsApp(
-      {String msg = '',
-      String imagePath = '',
-      FileType? fileType = FileType.image}) async {
+      {String msg = '', String imagePath = '', FileType? fileType = FileType.image}) async {
     final Map<String, dynamic> arguments = <String, dynamic>{};
     arguments.putIfAbsent('msg', () => msg);
     arguments.putIfAbsent('url', () => imagePath);
@@ -54,12 +54,25 @@ class FlutterShareMe {
 
     String? result;
     try {
-      result = await _channel.invokeMethod<String>(
-          _methodWhatsAppPersonal, arguments);
+      result = await _channel.invokeMethod<String>(_methodWhatsAppPersonal, arguments);
     } catch (e) {
       return e.toString();
     }
 
+    return result;
+  }
+  ///share to Telegram
+  /// [msg] message text you want on telegram
+  Future<String?> shareToTelegram(
+      {required String msg}) async {
+    final Map<String, dynamic> arguments = <String, dynamic>{};
+    arguments.putIfAbsent('msg', () => msg);
+    String? result;
+    try {
+      result = await _channel.invokeMethod<String>(_methodTelegramShare, arguments);
+    } catch (e) {
+      return e.toString();
+    }
     return result;
   }
 
@@ -67,16 +80,14 @@ class FlutterShareMe {
   ///[imagePath] is local image
   /// For ios
   /// If include image then text params will be ingored as there is no current way in IOS share both at the same.
-  Future<String?> shareToWhatsApp4Biz(
-      {String msg = '', String? imagePath = ''}) async {
+  Future<String?> shareToWhatsApp4Biz({String msg = '', String? imagePath = ''}) async {
     final Map<String, dynamic> arguments = <String, dynamic>{};
 
     arguments.putIfAbsent('msg', () => msg);
     arguments.putIfAbsent('url', () => imagePath);
     String? result;
     try {
-      result = await _channel.invokeMethod<String>(
-          _methodWhatsAppBusiness, arguments);
+      result = await _channel.invokeMethod<String>(_methodWhatsAppBusiness, arguments);
     } catch (e) {
       return 'false';
     }
@@ -85,8 +96,7 @@ class FlutterShareMe {
   }
 
   ///share to facebook
-  Future<String?> shareToFacebook(
-      {required String msg, String url = ''}) async {
+  Future<String?> shareToFacebook({required String msg, String url = ''}) async {
     final Map<String, dynamic> arguments = <String, dynamic>{};
     arguments.putIfAbsent('msg', () => msg);
     arguments.putIfAbsent('url', () => url);
@@ -118,8 +128,7 @@ class FlutterShareMe {
   Future<String?> shareToSystem({required String msg}) async {
     String? result;
     try {
-      result =
-          await _channel.invokeMethod<String>(_methodSystemShare, {'msg': msg});
+      result = await _channel.invokeMethod<String>(_methodSystemShare, {'msg': msg});
     } catch (e) {
       return 'false';
     }
@@ -127,8 +136,7 @@ class FlutterShareMe {
   }
 
   ///share file to instagram
-  Future<String?> shareToInstagram(
-      {required String imagePath }) async {
+  Future<String?> shareToInstagram({required String imagePath}) async {
     final Map<String, dynamic> arguments = <String, dynamic>{};
     arguments.putIfAbsent('url', () => imagePath);
     String? result;